+++ ElvUI_EltreumUI/Core/Options.lua
@@ -3667,30 +3667,20 @@ The Item Level shown on the Character Panel Skin uses code from Simple Item leve
 	ElvUI_EltreumUI.Options.args.changelog.args.added = ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = ACH:Description([[
-Added an option to disable the Unitframe Heal Prediction Texture override
-Added options to use different Custom Textures in Party and Raid Frames
-Added [name:eltruism:capital] tag
-Added a few missing shadows
+Added support for Season of Discovery Season 4
 ]], 3, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = ACH:Description([[
-Updated Focus Castbar Border size to use the Focus Border size
+Updated Item Level text on character panel to follow some of ElvUI's options
+Updated Background Alpha in the Character Panel Skin to not follow Crest Alpha
+Updated Character Panel Gradient and Class Colors to work separately
 ]], 5, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = ACH:Description([[
-Fixed an issue with instant loot where it could get stuck if a corpse was looted more than once in Cataclysm
-Fixed an issue where the Character Panel Camera settings were not working in Cataclysm Classic
-Fixed an issue where the EltruismDatatext texture would not hide with the datatext
-Fixed an issue where some items were not being pushed to the LootText
-Fixed an issue with the install due to a change in Shadow and Light
-Fixed an issue with the EltruismDatatext when it was loading
-Fixed an error with the Honor Datatext in Cataclysm Classic
-Fixed an issue with the DBM skin due to its update
-Fixed OmniCD skin and profile  after its update
-Fixed BigWigs profile after its update
-Fixed a leftover debug print
+Fixed a possible error when using /way command with text instead of zone id
+Fixed Combat Icons option missing
 ]], 7, "small", nil, nil, nil, nil, "full")
 
 	--[[