+++ Changelog.md
@@ -1,6 +1,13 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.1.8
+Added support for Season of Discovery Season 4  
+Updated Item Level text on character panel to follow some of ElvUI's options  
+Updated Background Alpha in the Character Panel Skin to not follow Crest Alpha  
+Updated Character Panel Gradient and Class Colors to work separately  
+Fixed a possible error when using /way command with text instead of zone id  
+Fixed Combat Icons option missing  
 ### 4.1.7.1
 Fixed a leftover debug print  
 ### 4.1.7